Remove fedora functionality from release-bot.
