+++ .s2i-dev/bin/run
@@ -17,8 +17,4 @@ export LD_PRELOAD=libnss_wrapper.so
 export NSS_WRAPPER_PASSWD="${HOME}/passwd"
 export NSS_WRAPPER_GROUP=/etc/group
 
-if [ -f "${HOME}/fedora.keytab" ]; then
-    release-bot --debug --configuration="${HOME}/conf.yaml" --keytab="${HOME}/fedora.keytab"
-else
-    release-bot --debug --configuration="${HOME}/conf.yaml"
-fi
+release-bot --debug --configuration="${HOME}/conf.yaml"
