+++ Dockerfile.dev
@@ -15,13 +15,12 @@ LABEL summary="Automated releasing from GitHub repositories" \
       io.openshift.s2i.scripts-url="$STI_SCRIPTS_URL" \
       usage="s2i build <CONFIGURATION-REPOSITORY> usercont/release-bot:dev <APP-NAME>"
 
-RUN dnf install -y fedpkg git krb5-workstation nss_wrapper
+RUN dnf install -y git nss_wrapper
 
 RUN mkdir -p ${HOME} && \
     useradd -u 1001 -r -g 0 -d ${HOME} -s /sbin/nologin \
     -c "Default Application User" default && \
-    chown -R 1001:0 /opt/app-root && \
-    sed 's/default_ccache_name/#default_ccache_name/g' /etc/krb5.conf > /etc/krb5.conf
+    chown -R 1001:0 /opt/app-root
 
 USER 1001
