+++ Dockerfile.test
@@ -1,6 +1,6 @@
 FROM registry.fedoraproject.org/fedora:29
 
-RUN dnf install -y fedpkg git make
+RUN dnf install -y git make
 
 ENV HOME=/home/test-user \
     PYTHONDONTWRITEBYTECODE=1
