--- tests/src/example.spec
@@ -1,15 +1,0 @@
-Name:           example
-Version:        0.0.1
-Release:        2%{?dist}
-Summary:        Summary
-
-License: 		GPLv3+
-URL:			www.example.com
-Source0:        www.example.com
-
-%description
-Example description
-
-%changelog
-* Wed Dec 06 2017 John Doe <jdoe@example.com> - 0.0.1-2
-- Initial package.
