--- tests/src/example_updated_changelog.spec
@@ -1,19 +1,0 @@
-Name:           example
-Version:        0.0.2
-Release:        1%{?dist}
-Summary:        Summary
-
-License: 		GPLv3+
-URL:			www.example.com
-Source0:        www.example.com
-
-%description
-Example description
-
-%changelog
-* Mon Dec 24 2018 John Doe <jdoe@example.com> 0.0.2-1
-- Changelog entry 1
-- Changelog entry 2
-
-* Wed Dec 06 2017 John Doe <jdoe@example.com> - 0.0.1-2
-- Initial package.
